Introduce Zig build system
Under the hood, Zig will use an embedded Clang to build our C++ source
files. However, Zig will automatically provide a lot of sane defaults
and other niceties, such as caching and soon its own linker. However,
this codebase is not up to snuff for a modern compiler. There's many,
many thousands of warnings and instances of Undefined Behavior that
make this build rather unstable and buggy. Currently, we must compile
with UBSan disabled and LTO disabled. As we identify and work through
the various problems in the code, we can start reenabling things and
eventually work towards mainlining this.

Note that this requires Zig 0.11.x to build which is not yet released.
However, https://ziglang.org/ provides builds from master branch. The
long-term plan is to vendor Zig ourselves, but let's get this working
first.